pyproject.toml
[tool.poetry]
name = "hr-bot"
version = "0.1.0"
description = ""
authors = ["Nick <nickphilomath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.2"
pymongo = "^4.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
